backend fyp fmd/requirements.txt
Flask==2.3.3
torch==2.1.0
torchvision==0.16.0
Pillow==10.0.0
